managesf: add platform policy management to the config repo
* Add a default policy.json file to the config repo
* Add a managesf config-update task
* Check policy.json is a valid json file in config-check
* Add a serverspec test to make sure the file is created


Related-To: #1497

+++ functestslib.sh
@@ -349,10 +349,25 @@ function run_bootstraps {
     eval $(ssh-agent)
     ssh-add ~/.ssh/id_rsa
     echo "$(date) ======= run_bootstraps" | tee -a ${ARTIFACTS_DIR}/bootstraps.log
+
+    echo "POLICY: prepare policy to support project creation by any user..."
+    ssh ${SF_HOST} python <<SCRIPT
+import json
+import os
+f = '/usr/local/share/sf-config-repo/policies/policy.json'
+if not os.path.isfile(f):
+    exit(0)
+j = json.load(file(f))
+j['managesf.project:create'] = 'rule:authenticated_api'
+j['managesf.project:delete'] = 'rule:authenticated_api'
+json.dump(j, file(f, 'w'), indent=1)
+SCRIPT
+
     ssh -A -tt ${SF_HOST} sfconfig.sh &> ${ARTIFACTS_DIR}/sfconfig.log \
         && echo "sfconfig.sh: SUCCESS"  \
         || { kill -9 $SSH_AGENT_PID; fail "sfconfig.sh failed" ${ARTIFACTS_DIR}/sfconfig.log; }
     kill -9 $SSH_AGENT_PID
+    rm /tmp/pol_change.py
     checkpoint "run_bootstraps"
     fetch_bootstraps_data
 }
@@ -460,6 +475,13 @@ function run_upgrade {
     echo "[+] Copying new version (${IMAGE_PATH}/ -> /var/lib/lxc/${INSTALL_SERVER}/rootfs/${IMAGE_PATH})"
     sudo mkdir -p /var/lib/lxc/${INSTALL_SERVER}/rootfs/${IMAGE_PATH}/ || fail "Could not copy ${SF_VER}"
     sudo rsync -a --delete ${IMAGE_PATH}/ /var/lib/lxc/${INSTALL_SERVER}/rootfs/${IMAGE_PATH}/ || fail "Could not copy ${SF_VER}"
+
+    echo "POLICY: prepare policy to support project creation by any user..."
+    ssh ${SF_HOST} "cd /root/config; if [ -f policies/policy.json ]; then sed -i policies/policy.json -e 's/rule:admin_api/rule:authenticated_api/'; git add policies; git commit -a -m 'test policy...'; git push git+ssh://sftests.com/config master; else exit 1; fi" || {
+        sudo sed -i /var/lib/lxc/${INSTALL_SERVER}/rootfs/${IMAGE_PATH}/usr/local/share/sf-config-repo/policies/policy.json \
+            -e "s/rule:admin_api/rule:authenticated_api/"
+    }
+
     echo "[+] Running upgrade"
     ssh ${SF_HOST} "cd software-factory; ./upgrade.sh" || fail "Upgrade failed" "/var/lib/lxc/${INSTALL_SERVER}/rootfs/var/log/upgrade-bootstrap.log"
     echo "[+] Update sf-bootstrap-data"
